löschen button in modulübersicht bei core Modulen ausgeblendet

+++ ulicms/admin/inc/modules.php
@@ -3,20 +3,20 @@ $acl = new ACL ();
 if (! $acl->hasPermission ( "list_packages" )) {
 	noperms ();
 } else {
-	
+
 	// @FIXME: Hartgecodete Texte in Sprachdateien auslagern.
 	// Das hier sollte am besten gleichzeitig mit dem Redesign der Paketverwaltung geschehen.
-	
+
 	if (isset ( $_POST ["truncate_installed_patches"] ) and $acl->hasPermission ( "patch_management" )) {
 		db_query ( "TRUNCATE TABLE " . tbname ( "installed_patches" ) );
 	}
-	
+
 	if ($acl->hasPermission ( "remove_packages" )) {
-		
+
 		// Modul deinstallieren
 		if (isset ( $_GET ["remove"] )) {
 			$remove = basename ( $_GET ["remove"] );
-			
+
 			$type = $_GET ["type"];
 			$uninstalled = uninstall_module ( $remove, $type );
 			if ($uninstalled) {
@@ -56,12 +56,12 @@ if (! $acl->hasPermission ( "list_packages" )) {
 				$controller = ControllerRegistry::get ( $main_class );
 			}
 			$module_has_admin_page = ((file_exists ( getModuleAdminFilePath ( $modules [$i] ) ) or file_exists ( getModuleAdminFilePath2 ( $modules [$i] ) ) or ($controller and method_exists ( $controller, "settings" ))) && ! faster_in_array ( $modules [$i], $disabledModules ));
-			
+
 			echo getModuleName ( $modules [$i] );
 			$version = getModuleMeta ( $modules [$i], "version" );
 			$source = getModuleMeta ( $modules [$i], "source" );
 			$color = null;
-			
+
 			if ($version != null) {
 				if ($source != "extend") {
 					$status = $pkg->checkForNewerVersionOfPackage ( $modules [$i] );
@@ -75,22 +75,22 @@ if (! $acl->hasPermission ( "list_packages" )) {
 						$color = "green";
 					}
 				}
-				
+
 				if ($color) {
 					echo '<span style="color: ' . $color . '">';
 				}
-				
+
 				echo " " . $version;
-				
+
 				if ($color) {
 					echo "</span>";
 				}
 			}
-			
+
 			echo "</strong>";
-			
+
 			echo "<div style=\"float:right\">";
-			
+
 			if ($module_has_admin_page) {
 				echo "<a style=\"font-size:0.8em;\" href=\"?action=module_settings&module=" . $modules [$i] . "\">";
 				$text = get_translation ( "settings" );
@@ -100,30 +100,30 @@ if (! $acl->hasPermission ( "list_packages" )) {
 				echo "[" . $text . "]";
 				echo "</a>";
 			}
-			
-			if ($acl->hasPermission ( "remove_packages" )) {
+
+			if ($acl->hasPermission ( "remove_packages" ) and $source != "core") {
 				echo " <a style=\"font-size:0.8em;\" href=\"?action=modules&remove=" . $modules [$i] . "&type=module\" onclick=\"return uninstallModule(this.href, '" . $modules [$i] . "');\">";
 				echo " [" . get_translation ( "delete" ) . "]";
 				echo "</a>";
 			}
-			
+
 			echo "</div>";
 			$noembed_file1 = getModulePath ( $modules [$i] ) . ".noembed";
 			$noembed_file2 = getModulePath ( $modules [$i] ) . "noembed.txt";
 			$embed_attrib = true;
-			
+
 			$meta_attr = getModuleMeta ( $modules [$i], "embed" );
 			if (! is_null ( $meta_attr ) and is_bool ( $meta_attr )) {
 				$embed_attrib = $meta_attr;
 			}
-			
+
 			echo "<br/>";
 			if (! file_exists ( $noembed_file1 ) and ! file_exists ( $noembed_file2 ) and $embed_attrib) {
 				$disabled = "";
 				if (faster_in_array ( $modules [$i], $disabledModules )) {
 					$disabled = "disabled";
 				}
-				
+
 				echo "<input type='text' value='[module=\"" . $modules [$i] . "\"]' readonly='readonly' " . $disabled . " onclick='this.focus(); this.select()'>";
 			} else {
 				echo "Kein Embed Modul";
@@ -151,13 +151,13 @@ if (! $acl->hasPermission ( "list_packages" )) {
 		echo "<ol>";
 		for($i = 0; $i < count ( $themes ); $i ++) {
 			echo "<li style=\"margin-bottom:20px;padding-bottom:10px;border-bottom:solid #cdcdcd 1px;\" id=\"dataset-theme-" . $themes [$i] . "\"><strong>";
-			
+
 			echo $themes [$i];
-			
+
 			$version = getThemeMeta ( $themes [$i], "version" );
 			$source = getThemeMeta ( $themes [$i], "source" );
 			$color = null;
-			
+
 			if ($version != null) {
 				if ($source != "extend") {
 					$status = $pkg->checkForNewerVersionOfPackage ( "theme-" . $themes [$i] );
@@ -171,22 +171,22 @@ if (! $acl->hasPermission ( "list_packages" )) {
 						$color = "green";
 					}
 				}
-				
+
 				if ($color) {
 					echo '<span style="color: ' . $color . '">';
 				}
-				
+
 				echo " " . $version;
-				
+
 				if ($color) {
 					echo "</span>";
 				}
 			}
-			
+
 			echo "</strong>";
-			
+
 			echo "<div style=\"float:right\">";
-			
+
 			if ($acl->hasPermission ( "remove_packages" ) and $themes [$i] != $ctheme) {
 				echo " <a style=\"font-size:0.8em;\" href=\"?action=modules&remove=" . $themes [$i] . "&type=theme\" onclick=\"return uninstallTheme(this.href, '" . $themes [$i] . "');\">";
 				echo " [" . get_translation ( "delete" ) . "]";
@@ -196,9 +196,9 @@ if (! $acl->hasPermission ( "list_packages" )) {
 				echo " [" . get_translation ( "delete" ) . "]";
 				echo "</a>";
 			}
-			
+
 			echo "</div>";
-			
+
 			"</li>";
 		}
 		echo "</ol>";
@@ -211,7 +211,7 @@ if (! $acl->hasPermission ( "list_packages" )) {
 <a name="installed_patches_a" id="installed_patches_a"></a>
 <p>
 	<strong><?php
-		
+
 		translate ( "installed_patches" );
 		?>
 </strong>
@@ -221,7 +221,7 @@ if (! $acl->hasPermission ( "list_packages" )) {
 		?>
 	</p>
 <?php
-		
+
 		if ($acl->hasPermission ( "upload_patches" )) {
 			?>
 <p>
@@ -238,7 +238,7 @@ if (! $acl->hasPermission ( "list_packages" )) {
 				echo "<li>" . htmlspecialchars ( $patch ) . "</li>";
 			}
 			echo "</ol>";
-			
+
 			?>
 <form id="truncate_installed_patches" action="index.php?action=modules"
 		method="post"
